fix: break circular import between Matrix and RC settings API modules
